requirements.txt
Cheetah==2.4.4
Landscape-Client==14.01
PAM==0.4.2
Pillow==2.3.0
PyYAML==3.10
Twisted-Core==13.2.0
Twisted-Names==13.2.0
Twisted-Web==13.2.0
apt-xapian-index==0.45
argparse==1.2.1
chardet==2.0.1
cloud-init==0.7.5
colorama==0.2.5
configobj==4.7.2
decorator==3.4.0
html5lib==0.999
jsonpatch==1.3
jsonpointer==1.0
nose==1.3.1
numpy==1.8.2
oauth==1.0.1
prettytable==0.7.2
pyOpenSSL==0.13
pycups==1.9.66
pycurl==7.19.3
pygobject==3.12.0
pyserial==2.6
pysmbc==1.0.14.1
python-apt==0.9.3.5
python-debian==0.1.21-nmu2ubuntu2
requests==2.2.1
scipy==0.13.3
six==1.8.0
ssh-import-id==3.21
stevedore==1.0.0
system-service==0.1.6
urllib3==1.7.1
virtualenv==1.11.6
virtualenv-clone==0.2.5
virtualenvwrapper==4.3.1
wsgiref==0.1.2
zope.interface==4.0.5
